# openapi/openapi.json
{
  "components": {
    "responses": {
      "Forbidden": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Authenticated, but the credential may not act on this resource. Note: cross-TENANT access returns 404, not 403, so the API is not an existence oracle."
      },
      "Unauthorized": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Missing, malformed or expired credential."
      }
    },
    "schemas": {
...
      "Error": {
        "properties": {
          "error": {
            "description": "Stable refusal kind (deny_* / invalid_credentials / …).",
            "type": "string"
          },
          "reason": {
            "description": "Single-line, secret-free human explanation.",
            "type": "string"
          }
        },
        "required": [
          "error",
          "reason"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "A SaaS session token (login) or a tenant API key. The tenant context is the credential's own organization unless the caller is a platform administrator.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Multi-tenant SaaS contract over the TASK 7A control plane. Trading truth (orders, fills, positions, risk, ledger, HA) remains owned by the TASK 1–6 engines and is exposed read-only to tenants. Billing/usage reads are served through the deterministic exports (`/api/saas/exports?kind=subscription|usage|...`); there are no separate plan-catalogue or usage endpoints.",
    "title": "Sniper Suite SaaS Control Plane",
    "version": "2.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/status": {
      "get": {
        "description": "Operator service status (engines, feeds, storage).",
        "operationId": "operator.status",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Status."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "tags": [
          "operator"
        ]
      }
    },
...
    "/api/tenant/limit-orders": {
      "get": {
        "description": "This organization's limit orders, newest first. Filter with `status`; page with `limit` (1-200) and `before` (RFC 3339 created_at, exclusive).",
        "operationId": "tenant.listLimitOrders",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "active",
                "triggered",
                "cancelled",
                "expired"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "A page of orders."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Invalid filter."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Storage unavailable. Never returns fabricated orders."
          }
        },
        "tags": [
          "tenant"
        ]
      },
      "post": {
        "description": "Create an ACTIVE limit order. Records intent only; execution is performed by the limit-order worker when the trigger is met.",
        "operationId": "tenant.createLimitOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Order created."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Validation failed (side, trigger, mint, price, amount or expiry)."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Storage unavailable."
          }
        },
        "tags": [
          "tenant"
        ]
      }
    },
    "/api/tenant/limit-orders/{id}/cancel": {
      "post": {
        "description": "Cancel an ACTIVE order that no worker is currently executing. Refusals are explicit: 409 when already terminal or being triggered, 404 when not found in this organization.",
        "operationId": "tenant.cancelLimitOrder",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Cancelled."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Not found in this organization."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Not active, or being triggered."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Storage unavailable."
          }
        },
        "tags": [
          "tenant"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "the control-plane deployment itself",
      "url": "/"
    }
  ],
...
}